👷 [+ci] Update build scripts to produce working binary
The updated dependency on `dbrownell_Common` caused an update in `inflect`. The new version of `inflect` introduced a dependency on `typeguard`. `typeguard` introduces code that uses `inspect`, which was failing when compiled with `cx_Freeze` because the python source was not found. This change ensure that inflect's source code is bundled with the binary, which then makes `typeguard` happy. Then, `inflect` is happy, which makes `dbrownell_Common` happy, which makes `FileBackup` work as expected when frozen with `cx_Freeze`.